+++ src/main.rs
@@ -1,4 +1,7 @@
-use ggez::*;
+// contains all the information on entities
+mod entities;
+mod ui;
+mod sprites;
 
 mod states;
 use states::StateMachine;
@@ -14,67 +17,16 @@ fn main() {
             .build()
             .unwrap();
 
-    // create player
-    let mut player = Player::new(ctx);
-    player.move_location(150f32, 150f32);
-	let hp = player.hp;
+    let mut state_machine = StateMachine::new();
+    
+    //state_machine.new_main_state(ctx);
+    //let state = &mut states::MainState::new(ctx);
 
-    // create blobs (ie enemies)
-    let mut blob = Vec::new();
-    blob.push(Blob::new(ctx, 250.0, 250.0));
-    blob.push(Blob::new(ctx, 250.0, 350.0));
-    blob.push(Blob::new(ctx, 250.0, 150.0));
-
-    // build level
-    let img = graphics::Image::new(ctx, "/testwalls.png").unwrap();
-    let mut lb = LevelBuilder::new(ctx, None);
-    lb.set_tile_image(
-        0usize,
-        &Sprite::new(&img, graphics::Rect::new(0f32, 0f32, 64f32, 64f32)).unwrap(),
-    )
-    .unwrap();
-    lb.set_tile_image(
-        1usize,
-        &Sprite::new(&img, graphics::Rect::new(64f32, 0f32, 64f32, 64f32)).unwrap(),
-    )
-    .unwrap();
-    lb.set_tile_image(
-        2usize,
-        &Sprite::new(&img, graphics::Rect::new(128f32, 0f32, 64f32, 64f32)).unwrap(),
-    )
-    .unwrap();
-    lb.set_tile_image(
-        3usize,
-        &Sprite::new(&img, graphics::Rect::new(192f32, 0f32, 64f32, 64f32)).unwrap(),
-    )
-    .unwrap();
-    let level = lb.sample3();
-
-    // demo sprites
-    let img = graphics::Image::new(ctx, "/dapper-skeleton-sheet.png").unwrap();
-    let sprite = Sprite::new(&img, graphics::Rect::new(0f32, 128f32, 64f32, 64f32)).unwrap();
-    let animated = AnimatedBuilder::new(&img)
-        .create_animated(graphics::Rect::new(0f32, 320f32, 64f32, 64f32), 6usize)
-        .unwrap();
-
-    // create state
-    let state = &mut MainState {
-        level,
-        blob,
-        player,
-		ui: UI::new(ctx, "Adventurer".to_string(), hp),
-        sprite,
-        animated,
-        rotation: 0f32,
-    };
-	
-	let state = &mut MainMenuState {
-		ui: UI::new(ctx, "Adventurer".to_string(), 10.0),
-	};
+    state_machine.run(ctx, event_loop);
 
     // start game loop
-    match ggez::event::run(ctx, event_loop, state) {
-        Ok(_) => println!("Exiting Game."),
-        Err(e) => println!("Run event loop broke! {}", e),
-    }
+    //match ggez::event::run(ctx, event_loop, &mut state_machine.main_state.unwrap()) {
+    //    Ok(_) => println!("Exiting Game."),
+    //    Err(e) => println!("Run event loop broke! {}", e),
+    //}
 }
